Tidy up of trace spacing. Added larger pads to NFC chip.
Initial design had irregular spacing between traces on top and bottom layer. With this revision the lower copper was used for the spacing on the top layer, and vice versa. In this way a very even track spacing was achieved which cut down on overall size of PCB fractionally.
From the first batch initial success was achieved on 20 of 24 tags. Of the remaining 4 2 were fixed by touching solder to the caps.
The remaining issues were with the NXP chip where I could not get heat to bottom for reheat soldering. Slightly larger pads were added wxtending out from under the IC outline which will also allow accepting a little more solder paste.